+++ data.tf
@@ -1,13 +1,7 @@
 data "aws_region" "current" {}
 
-data "aws_ami" "eks_worker" {
-  filter {
-    name   = "name"
-    values = ["eks-worker-*"]
-  }
-
-  most_recent = true
-  owners      = ["602401143452"] # Amazon
+data "http" "workstation_external_ip" {
+  url = "http://icanhazip.com"
 }
 
 data "aws_iam_policy_document" "workers_assume_role_policy" {
@@ -25,6 +19,16 @@ data "aws_iam_policy_document" "workers_assume_role_policy" {
   }
 }
 
+data "aws_ami" "eks_worker" {
+  filter {
+    name   = "name"
+    values = ["eks-worker-*"]
+  }
+
+  most_recent = true
+  owners      = ["602401143452"] # Amazon
+}
+
 data "aws_iam_policy_document" "cluster_assume_role_policy" {
   statement {
     sid = "EKSClusterAssumeRole"
@@ -37,19 +41,6 @@ data "aws_iam_policy_document" "cluster_assume_role_policy" {
       type        = "Service"
       identifiers = ["eks.amazonaws.com"]
     }
-  }
-}
-
-data template_file userdata {
-  template = "${file("${path.module}/templates/userdata.sh.tpl")}"
-
-  vars {
-    region              = "${data.aws_region.current.name}"
-    max_pod_count       = "${lookup(local.max_pod_per_node, var.workers_instance_type)}"
-    cluster_name        = "${var.cluster_name}"
-    endpoint            = "${aws_eks_cluster.this.endpoint}"
-    cluster_auth_base64 = "${aws_eks_cluster.this.certificate_authority.0.data}"
-    additional_userdata = "${var.additional_userdata}"
   }
 }
 
@@ -72,7 +63,16 @@ data template_file config_map_aws_auth {
   }
 }
 
-module "ebs_optimized" {
-  source        = "./modules/tf_util_ebs_optimized"
-  instance_type = "${var.workers_instance_type}"
+data template_file userdata {
+  template = "${file("${path.module}/templates/userdata.sh.tpl")}"
+  count    = "${length(var.worker_groups)}"
+
+  vars {
+    region              = "${data.aws_region.current.name}"
+    cluster_name        = "${var.cluster_name}"
+    endpoint            = "${aws_eks_cluster.this.endpoint}"
+    cluster_auth_base64 = "${aws_eks_cluster.this.certificate_authority.0.data}"
+    max_pod_count       = "${lookup(local.max_pod_per_node, lookup(var.worker_groups[count.index], "instance_type", lookup(var.workers_group_defaults, "instance_type")))}"
+    additional_userdata = "${lookup(var.worker_groups[count.index], "additional_userdata",lookup(var.workers_group_defaults, "additional_userdata"))}"
+  }
 }