+++ outputs.tf
@@ -1,15 +1,5 @@
-output "config_map_aws_auth" {
-  description = "A kubernetes configuration to authenticate to this cluster."
-  value       = "${data.template_file.config_map_aws_auth.rendered}"
-}
-
-output "kubeconfig" {
-  description = "kubectl config file contents for this cluster."
-  value       = "${data.template_file.kubeconfig.rendered}"
-}
-
 output "cluster_id" {
-  description = "The name/id of the cluster."
+  description = "The name/id of the EKS cluster."
   value       = "${aws_eks_cluster.this.id}"
 }
 
@@ -20,21 +10,41 @@ output "cluster_id" {
 # }
 
 output "cluster_certificate_authority_data" {
-  description = "Nested attribute containing certificate-authority-data for your cluster. Tis is the base64 encoded certificate data required to communicate with your cluster."
+  description = "Nested attribute containing certificate-authority-data for your cluster. This is the base64 encoded certificate data required to communicate with your cluster."
   value       = "${aws_eks_cluster.this.certificate_authority.0.data}"
 }
 
 output "cluster_endpoint" {
-  description = "The endpoint for your Kubernetes API server."
+  description = "The endpoint for your EKS Kubernetes API."
   value       = "${aws_eks_cluster.this.endpoint}"
 }
 
 output "cluster_version" {
-  description = "The Kubernetes server version for the cluster."
+  description = "The Kubernetes server version for the EKS cluster."
   value       = "${aws_eks_cluster.this.version}"
 }
 
-output "cluster_security_group_ids" {
-  description = "description"
-  value       = "${aws_eks_cluster.this.vpc_config.0.security_group_ids}"
+output "cluster_security_group_id" {
+  description = "Security group ID attached to the EKS cluster."
+  value       = "${local.cluster_security_group_id}"
+}
+
+output "config_map_aws_auth" {
+  description = "A kubernetes configuration to authenticate to this EKS cluster."
+  value       = "${data.template_file.config_map_aws_auth.rendered}"
+}
+
+output "kubeconfig" {
+  description = "kubectl config file contents for this EKS cluster."
+  value       = "${data.template_file.kubeconfig.rendered}"
+}
+
+output "workers_asg_arns" {
+  description = "IDs of the autoscaling groups containing workers."
+  value       = "${aws_autoscaling_group.workers.*.arn}"
+}
+
+output "worker_security_group_id" {
+  description = "Security group ID attached to the EKS workers."
+  value       = "${local.worker_security_group_id}"
 }