+++ CHANGELOG.md
@@ -5,22 +5,36 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## [[v1.0.0](https://github.com/terraform-aws-modules/terraform-aws-eks/compare/v0.2.0...v1.0.0)] - 2018-06-11]
+
+### Added
+
+- security group id can be provided for either/both of the cluster and the workers. If not provided, security groups will be created with sufficient rules to allow cluster-worker communication. - kudos to @tanmng on the idea ⭐
+- outputs of security group ids and worker ASG arns added for working with these resources outside the module.
+
+### Changed
+
+- Worker build out refactored to allow multiple autoscaling groups each having differing specs. If none are given, a single ASG is created with a set of sane defaults - big thanks to @kppullin 🥨
+
 ## [[v0.2.0](https://github.com/terraform-aws-modules/terraform-aws-eks/compare/v0.1.1...v0.2.0)] - 2018-06-08]
+
+### Added
+
+- ability to specify extra userdata code to execute following kubelet services start.
+- EBS optimization used whenever possible for the given instance type.
+- When `configure_kubectl_session` is set to true the current shell will be configured to talk to the kubernetes cluster using config files output from the module.
 
 ### Changed
 
 - files rendered from dedicated templates to separate out raw code and config from `hcl`
 - `workers_ami_id` is now made optional. If not specified, the module will source the latest AWS supported EKS AMI instead.
-- added ability to specify extra userdata code to execute after the second to configure and start kube services.
-- When `configure_kubectl_session` is set to true the current shell will be configured to talk to the kubernetes cluster using config files output from the module.
-- EBS optimization used whenever possible for the given instance type.
 
 ## [[v0.1.1](https://github.com/terraform-aws-modules/terraform-aws-eks/compare/v0.1.0...v0.1.1)] - 2018-06-07]
 
 ### Changed
 
-- pre-commit hooks fixed and working.
-- made progress on CI, advancing the build to the final `kitchen test` stage before failing.
+- Pre-commit hooks fixed and working.
+- Made progress on CI, advancing the build to the final `kitchen test` stage before failing.
 
 ## [v0.1.0] - 2018-06-07